Moved report mock data to mockData.js and updated test for NswgReporter to use async/await instead of Promise
