+++ __mocks__/mockData.js
@@ -51,7 +51,65 @@ const reportUserData = {
   } 
 }
 
+const report = { 
+  data: { 
+    id: '123456',
+    type: 'report',
+    attributes: { 
+      title : "overview",
+      state: 'resolved',
+      created_at: '2018-07-13T10:04:52.180Z',
+      triaged_at: null,
+      closed_at: null,
+      last_reporter_activity_at: null,
+      first_program_activity_at: null,
+      last_program_activity_at: null,
+      bounty_awarded_at: null,
+      swag_awarded_at: null,
+      disclosed_at: '2018-08-22T07:48:35.622Z',
+      last_public_activity_at: null,
+      last_activity_at: null,
+      cve_ids: null 
+    },
+    relationships: { 
+      reporter: [Object],
+      program: [Object],
+      swag: [Object],
+      attachments: [Object],
+      activities: [Object],
+      bounties: [Object],
+      summaries: [Object] } },
+ reporter: { 
+    id: 123456,
+    username: 'asdfghjkjl',
+    name: 'asdfghjkl',
+    bio: null,
+    website: null,
+    location: null,
+    created_at: '2018-07-04T13:30:25.133Z',
+    url: 'https://hackerone.com/asdfghjkl',
+    anc_triager: false,
+    hackerone_triager: false,
+    profile_picture_urls: { 
+      small: '/assets/avatars/default-12345.png',
+      medium: '/assets/avatars/default-123456.png',
+      xtralarge: '/assets/avatars/default-1234567.png' },
+    disabled: false,
+    report_count: 6,
+    target_count: 1,
+    reputation: 142,
+    rank: 2630,
+    signal: 7,
+    impact: null,
+    signal_percentile: 96,
+    impact_percentile: null,
+    total_mentees: 0,
+    mentor: null } 
+  }
+
+
 module.exports = {
+  report,
   reportData, 
   reportUserData
 }
